# examples/simple/openapi.json
{
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header"
      }
    }
  },
  "openapi": "3.1.0",
  "info": {
    "title": "Simple Example",
    "version": "1.0.0"
  },
  "paths": {
    "/widgets": {
      "get": {
        "operationId": "WidgetsController.getAllWidgets",
        "responses": {
          "200": {
            "description": "All widgets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string",
                        "minLength": 1
                      },
                      "disposition": {
                        "type": "string",
                        "enum": [
                          "happy",
                          "sad"
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "disposition"
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "disposition",
            "description": "Filter by disposition",
            "schema": {
              "type": "string",
              "enum": [
                "happy",
                "sad"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "description": "Limit the number of results",
            "schema": {
              "type": "number",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "offset",
            "description": "Offset the results",
            "schema": {
              "type": "number",
              "minimum": 0
            }
          }
        ],
        "description": "Get all widgets",
        "tags": [
          "widget"
        ]
      },
      "post": {
        "operationId": "WidgetsController.createWidget",
        "responses": {
          "201": {
            "description": "The created widget",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "minLength": 1
                },
                "description": "The location of the created widget",
                "required": true
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string",
                      "minLength": 1
                    },
                    "disposition": {
                      "type": "string",
                      "enum": [
                        "happy",
                        "sad"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "disposition"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid widget",
            "content": {}
          }
        },
        "requestBody": {
          "required": true,
          "description": "The content of the widget",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "disposition": {
                    "type": "string",
                    "enum": [
                      "happy",
                      "sad"
                    ]
                  }
                },
                "required": [
                  "name",
                  "disposition"
                ]
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [
              "write:widgets"
            ]
          }
        ],
        "description": "Create a widget",
        "tags": [
          "widget"
        ]
      }
    },
    "/widgets/{id}": {
      "get": {
        "operationId": "WidgetsController.getWidgetById",
        "responses": {
          "200": {
            "description": "The widget",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string",
                      "minLength": 1
                    },
                    "disposition": {
                      "type": "string",
                      "enum": [
                        "happy",
                        "sad"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "disposition"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Widget not found",
            "content": {}
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the widget",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "description": "Get a widget by ID",
        "tags": [
          "widget"
        ]
      }
    }
  }
}